+++ src/main/kotlin/com/github/niclasvaneyk/laravelmake/common/collection/map.kt
@@ -1,0 +1,5 @@
+package com.github.niclasvaneyk.laravelmake.common.collection
+
+val Map<String, Any>.humanReadableKeys: String get() {
+    return keys.joinToString(", ") { "'$it'" }
+}