+++ src/main/kotlin/com/github/niclasvaneyk/laravelmake/common/laravel/introspection/CommandRunInfo.kt
@@ -1,0 +1,7 @@
+package com.github.niclasvaneyk.laravelmake.common.laravel.introspection
+
+class CommandRunInfo(
+    val namespace: String,
+    val command: String?,
+    val options: List<String> = emptyList()
+)