+++ src/main/kotlin/com/github/niclasvaneyk/laravelmake/common/php/run/ArbitraryPhpRunner.kt
@@ -1,0 +1,9 @@
+package com.github.niclasvaneyk.laravelmake.common.php.run
+
+import com.github.niclasvaneyk.laravelmake.common.laravel.Artisan
+
+class ArbitraryPhpRunner(private val artisan: Artisan) {
+    fun run(php: String): PHPRunner.Result {
+        return artisan.command("tinker", parameters = listOf("--execute", php))
+    }
+}