+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 ### Added
 
+- Settings! Check them out under "Languages & Frameworks > PHP > Laravel Make"
+- Setting for disabling the "Autoconfigure Laravel Sail" notification
+- Setting for hiding/disabling/keeping irrelevant artisan:make entries in the "File > New > Laravel" context menu 
+
 ### Changed
 
 ### Fixed
@@ -67,7 +71,7 @@
 
 ### Fixed
 
-- PhpStorm does not crash anymore if you open a non-Laravel project. Thank you Jeremy for reporting this!
+- PhpStorm does not crash anymore if you open a non-Laravel project. Thank you, Jeremy, for reporting this!
 
 ## [2.3.0]
 
@@ -118,7 +122,7 @@
 ## [2.0.0]
 
 - You can pass flags to the commands (#3), just as you would do on the command line
-- The artisan binary now gets executed using the project interpreter, wich has the following impact
+- The artisan binary now gets executed using the project interpreter, which has the following impact
     - The plugin now requires a paid IDE, because it now depends on the PHP-Plugin
     - Remote Interpreters are supported (#1), so now Docker containers or SSH can be used to execute the artisan commands
 - Updated the plugin infrastructure to use Gradle (should have no user-facing changes)